# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy API",
        "description": "RESTful API theo chuẩn JSON:API",
        "contact": {
            "email": "dongockhanh2003.utt@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main Server"
        }
    ],
    "paths": {
        "/v1/store/account/address/add": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Thêm địa chỉ mới cho người dùng",
                "description": "Thêm một địa chỉ mới vào danh sách địa chỉ của người dùng đã đăng nhập",
                "operationId": "addAddress",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address_line_1",
                                    "city",
                                    "country",
                                    "postal_code"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Tên người nhận",
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "phone": {
                                        "description": "Số điện thoại",
                                        "type": "string",
                                        "example": "0901234567"
                                    },
                                    "address_line_1": {
                                        "description": "Địa chỉ dòng 1",
                                        "type": "string",
                                        "example": "123 Đường Nguyễn Huệ"
                                    },
                                    "address_line_2": {
                                        "description": "Địa chỉ dòng 2",
                                        "type": "string",
                                        "example": "Phường Bến Nghé"
                                    },
                                    "city": {
                                        "description": "Quận/Huyện",
                                        "type": "string",
                                        "example": "Quận 1"
                                    },
                                    "state": {
                                        "description": "Tỉnh/Thành phố",
                                        "type": "string",
                                        "example": "TP Hồ Chí Minh"
                                    },
                                    "country": {
                                        "description": "Quốc gia",
                                        "type": "string",
                                        "example": "Việt Nam"
                                    },
                                    "postal_code": {
                                        "description": "Mã bưu điện",
                                        "type": "string",
                                        "example": "700000"
                                    },
                                    "is_default": {
                                        "description": "Đặt làm địa chỉ mặc định",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Địa chỉ đã được thêm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nguyễn Văn A"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "0901234567"
                                                },
                                                "address_line_1": {
                                                    "type": "string",
                                                    "example": "123 Đường Nguyễn Huệ"
                                                },
                                                "address_line_2": {
                                                    "type": "string",
                                                    "example": "Phường Bến Nghé"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Quận 1"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "TP Hồ Chí Minh"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "Việt Nam"
                                                },
                                                "postal_code": {
                                                    "type": "string",
                                                    "example": "700000"
                                                },
                                                "is_default": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Địa chỉ đã được thêm thành công"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "vi_VN"
                                        },
                                        "error": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Địa chỉ đã tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Địa chỉ đã tồn tại"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "vi_VN"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "locale": {
                                            "type": "string",
                                            "example": "vi_VN"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Xác thực bằng JWT Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}